backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
numpy
opencv-python
pillow
mlflow
matplotlib
scikit-learn
python-dotenv
tqdm
